+++ inst/src/lib/libraries.R
@@ -17,6 +17,7 @@ if (!require("tidyr")) install.packages("tidyr")
 if (!require("scales")) install.packages("scales")
 if (!require("broom")) install.packages("broom")
 if (!require("plotly")) install.packages("plotly")
+if (!require("Rtsne")) install.packages("Rtsne")
 if (!require("rmarkdown")) install.packages("rmarkdown")
 
 ## Pathway analysis